avatar-gamer/backend/requirements.txt
django
djangorestframework
psycopg2-binary
python-dotenv
django-cors-headers
djangorestframework-simplejwt
pillow
requests
python-socketio
uvicorn
redis
aioredis
